requirements.txt
pypdf==4.0.1
scikit-learn>=1.4.0
requests==2.31.0
pandas>=2.2.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
python-dotenv
pypdf
